Exercises/Chapter 24/24.5 Hashing and Log In/requirements.txt
autopep8==2.0.0
bcrypt==4.0.1
blinker==1.5
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycodestyle==2.9.1
SQLAlchemy==1.4.42
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
